redirected properties page to root directory
