+++ src/pages/index.js
@@ -1,22 +1,30 @@
-import Head from "next/head";
-import { useAuthUser } from "@/firebase/auth";
-import RoommatesFilter from "../components/roommatesFilter";
+import * as React from 'react';
+import Grid from '@mui/material/Grid';
+import PropertyFilter from '@/components/PropertyFilter/PropertyFilter';
+import PropertyCard from '@/components/PropertyCard';
+import { getAllProperties } from '@/firebase/propertyDb';
 
-export default function Home() {
-	const [user, loading] = useAuthUser();
+export default function ComplexGrid(props) {
+	return (
+		<div style={{ display: 'flex', padding: '10px 0px 0px 20px' }}>
+			<PropertyFilter />
+			<Grid container direction="row" spacing={0} sx={{ px: 5 }}>
+				{props.properties.map((property, i) => {
+					return (
+						<Grid item display="flex" justifyContent="center" alignItems="center" key={i} sm>
+							<PropertyCard property={property} />
+						</Grid>
+					);
+				})}
+			</Grid>
+		</div>
+	);
+}
 
-  if (loading) {
-    return <h1>Loading...</h1>;
-  }
-  return (
-    <>
-      <RoommatesFilter />
-      <h1>{user ? user.displayName : "Need to login"}</h1>
-      <Head>
-        <title>Create Next App</title>
-        <meta name="description" content="Generated by create next app" />
-        <meta name="viewport" content="width=device-width, initial-scale=1" />
-      </Head>
-    </>
-  );
+export async function getStaticProps() {
+	const properties = await getAllProperties();
+	return {
+		props: { properties }, // will be passed to the page component as props
+		revalidate: 600, // update cache every 10 minutes
+	};
 }
